Graphite template test cases for enquos
Adds a method to print the Graphite template matcher parse tree to help
with debugging cases where the template does not match incoming metrics.

Additionally, adds unit tests to easily iterate over our configuration
and known metrics.

The `TestFilterMatchEnquos.templates` list needs to be adjusted for the
latest production configuration and test cases added to the
`TestFilterMatchEnquos.test_cases` list as needed.

Run the tests with

  cd $GOPATH/src/github.com/ABNVCO/influxdb
  go build ./... && DEBUG_GRAPHITE=1 go test -run=Enquos ./services/graphite -v